+++ CHANGELOG.md
@@ -19,6 +19,6 @@
 - Moved `csc_row_normalize`,  `numpy_rebuild_nnz_rowcol`, `set_to_zeroes`, `set_to_ones`, `find_spectral_gap`, ` remove_nnz_rowcol` from `temporal_network.py` to `utils.py`. 
 - Change the name of `Tcsr` to just `T` in function `set_to_zeroes` in `utils.py`
 - Moved `numpy_rebuild_nnz_rowcol` from `utils.py` to `stochmat` package.
-- Added `num_active_edges`, `num_active_nodes` as methods of `ContTempNetwork` class in the `temporal_network.py`. 
+- Added `num_active_events`, `num_active_nodes` as methods of `ContTempNetwork` class in the `temporal_network.py`. 
 They compute the number of active edges and nodes within a range of the t_start and t_end.